feat: 实现财务/出纳模块核心功能
- 新增 CASHIER 角色和 SETTLEMENT_AUDIT 权限
- 更新 settlement-manage 云函数：
  - auditSettlement: 实时计算扣款逻辑
  - completePayment: 出纳确认付款
  - getCashierStats: 出纳统计接口
  - previewDeduction: 扣款预览接口
- 更新 acquisition-manage: 结算单使用新的三阶段状态
- 首页添加出纳专属视图
- 更新权限配置和导航栏

+++ miniprogram/pages/operations/seed-add/index.ts
@@ -27,8 +27,8 @@ Page({
     // 表单数据
     form: {
       distributeTime: '',     // 发放时间
-      quantity: '',           // 发放数量（株）
-      unitPrice: '',          // 单价（元/株）
+      quantity: '',           // 发放数量（万株）
+      unitPrice: '',          // 单价（元/万株）
       distributedArea: '',    // 本次发放面积（亩）
       receiverName: '',       // 领取人
       receiveLocation: '',    // 领取地点
@@ -524,7 +524,7 @@ Page({
   // ==================== 表单输入 ====================
 
   /**
-   * 输入发放数量（株）
+   * 输入发放数量（万株）
    */
   onQuantityInput(e: WechatMiniprogram.CustomEvent) {
     // 只允许整数
@@ -536,7 +536,7 @@ Page({
   },
 
   /**
-   * 输入单价（元/株）
+   * 输入单价（元/万株）
    */
   onUnitPriceInput(e: WechatMiniprogram.CustomEvent) {
     let value = e.detail.value.replace(/[^\d.]/g, '');
@@ -605,11 +605,11 @@ Page({
 
   /**
    * 计算已发放金额
-   * 已发放金额 = 发放数量（株）× 单价（元/株）
+   * 已发放金额 = 发放数量（万株）× 单价（元/万株）
    */
   calculateAmount() {
     const { quantity, unitPrice } = this.data.form;
-    const qty = parseInt(quantity) || 0;  // 株数为整数
+    const qty = parseFloat(quantity) || 0;  // 万株（可为小数）
     const price = parseFloat(unitPrice) || 0;
     const amount = (qty * price).toFixed(2);
